refactored code to make it look cleaner
