+++ backend/database/database.js
@@ -1,0 +1,9 @@
+var mysql = require("mysql");
+var connection = mysql.createConnection({
+  host: "localhost",
+  user: "root",
+  password: "",
+  database: "ljps_db",
+});
+
+export default connection
