+++ backend/main/server.js
@@ -6,7 +6,8 @@ const bodyParser = require('body-parser');
 const app = express();  
 
 
-const api = require('./api');
+const RolesRouter = require('./routes/roles/RolesRouter');
+const SkillsRouter = require('./routes/skills/SkillsRouter')
 
 
 app.use(cors())
@@ -15,10 +16,10 @@ app.use(bodyParser.urlencoded({ extended: false }));
 app.use((req, res, next)=>{
     console.log(`${req.method} - ${req.url}`)
     
-    
     next()
 })
-app.use('/api', api);
+app.use('/api/roles', RolesRouter);
+app.use('/api/skills', SkillsRouter)
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
